fix(form-builder): add missing presence context
