+++ packages/@sanity/form-builder/src/__workshop__/_common/index.ts
@@ -1,0 +1,6 @@
+import {FilterFieldInput} from './FilterFieldInput'
+import {FormBuilderTester} from './FormBuilderTester'
+import {TypeTester} from './TypeTester'
+import {FormDebugger} from './FormDebugger'
+
+export {FormDebugger, FilterFieldInput, FormBuilderTester, TypeTester}
